last redact

+++ TransportDll/Air_carpet.cpp
@@ -4,8 +4,12 @@ Air_carpet::Air_carpet() :AirTrpt(10, 5) {
 	name = "Ковер-самолет";
 }
 
-double Air_carpet::RaceTime(Transport* Trnsp, double dist) {
+double Air_carpet::RaceTime(Transport* Trnsp, int dist) {
 	double time;
+	
+	//switch (dist) {
+	//case 0 ... 1000:
+	//}
 	if (dist < 1000) {
 		time = dist / speed;
 		return time;
